+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2019-08-22
+### Changed
+- Use symfony events
+- Changed command, event and handler classes and interfaces
+- Removed MessageBus from Handlers, throw CommandValidationException instead
+- Handlers can now be services (they must be public)
+### Removed
+- Removed abstract Handler class
+
 ## [1.1.3] - 2019-08-06
 ### Changed
 - Improved `$commandBus->execute()` convenience method